Split ShapeHelpers into shapes, paint and modifiers
Thirteen builders in one file, and they are three unrelated things: the shape
primitives (rect, ellipse, path, group), the paint applied to them (fill,
gradient, stroke, dashes), and the two items that modify a group rather than
draw anything (trim path, repeater). One file each.

Styles dumped before and after: unchanged.
